# docs/contracts/api-spec.yaml
openapi: "3.0.3"
info:
  title: 租车应用 API
  version: "2.0"
  description: 租车小程序 + PC 管理端 RESTful API
servers:
  - url: /api/v1
    description: v1 API
security:
  - BearerAuth: []
paths:
...
  /vehicles:
    get:
      summary: 车辆列表
      tags: [Vehicle]
      security: []
      parameters:
        - $ref: '#/components/parameters/min_price'
        - $ref: '#/components/parameters/max_price'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/page_size'
      responses:
        '200':
          description: 车辆列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleListResponse'
  /vehicles/{id}:
    get:
      summary: 车辆详情
      tags: [Vehicle]
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 车辆详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/VehicleDetail'
        '4004':
          description: 车辆不存在

  # ==================== 订单模块（用户端） ====================
...
  /admin/vehicles/{id}:
    put:
      summary: 更新车辆
      tags: [Admin Vehicle]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleRequest'
      responses:
        '200':
          description: 更新成功
        '4004':
          description: 车辆不存在
    delete:
      summary: 删除车辆（软删除）
      tags: [Admin Vehicle]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 删除成功
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    min_price:
      name: min_price
      in: query
      required: false
      schema:
        type: number
        format: float
      description: 最低日租金
    max_price:
      name: max_price
      in: query
      required: false
      schema:
        type: number
        format: float
      description: 最高日租金
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
      description: 页码
    page_size:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        default: 20
      description: 每页数量
  schemas:
...
    ApiResponse:
      type: object
      required: [code, data, message]
      properties:
        code:
          type: integer
          description: 错误码，0 表示成功
        data:
          type: object
          nullable: true
          description: 响应数据
        message:
          type: string
          description: 响应消息

    # ===== 认证相关 =====
...
    VehicleListItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        brand:
          type: string
        seats:
          type: integer
        transmission:
          type: string
          enum: [auto, manual]
        cover_image:
          type: string
          description: 封面图，取自 images[0]
        weekday_price:
          type: string
          format: decimal
          description: 工作日日租金（string 类型）
        weekend_price:
          type: string
          format: decimal
          description: 周末日租金（string 类型）
    VehicleDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        brand:
          type: string
        seats:
          type: integer
        transmission:
          type: string
          enum: [auto, manual]
        description:
          type: string
        images:
          type: array
          items:
            type: string
            format: uri
        weekday_price:
          type: string
          format: decimal
        weekend_price:
          type: string
          format: decimal
        holiday_price:
          type: string
          format: decimal
          nullable: true
        tags:
          type: array
          items:
            type: string
    VehicleListResponse:
      type: object
      properties:
        code:
          type: integer
        data:
          type: object
          properties:
            total:
              type: integer
            items:
              type: array
              items:
                $ref: '#/components/schemas/VehicleListItem'

    # ===== 订单相关 =====
...
    CreateVehicleRequest:
      type: object
      required: [name, brand, seats, transmission, weekday_price, weekend_price]
      properties:
        name:
          type: string
        brand:
          type: string
        seats:
          type: integer
        transmission:
          type: string
          enum: [auto, manual]
        description:
          type: string
        images:
          type: array
          items:
            type: string
            format: uri
        weekday_price:
          type: string
          format: decimal
        weekend_price:
          type: string
          format: decimal
        holiday_price:
          type: string
          format: decimal
          nullable: true
          description: 不填时默认使用 weekend_price

    # ===== 价格设置 =====
...